// src/api/orval/model/geoPointAggDto.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */

export interface GeoPointAggDto {
  /** @nullable */
  country?: string | null;
  /** @nullable */
  region?: string | null;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  sessionsCount?: number;
  totalBytesIn?: number;
  totalBytesOut?: number;
}